+++ src/dummies/b-dummy/interface.ts
@@ -18,7 +18,7 @@ import type updateOn from 'core/component/directives/update-on/engines';
 import type iObserveDOM from 'traits/i-observe-dom/i-observe-dom';
 
 import type InMemoryRouterEngine from 'core/router/engines/in-memory';
-import type HistoryApiRouterEngine from 'core/router/engines/browser.history';
+import type HistoryApiRouterEngine from 'core/router/engines/browser-history';
 
 export interface Directives {
 	imageFactory: typeof imageLoaderFactory;