+++ src/super/i-static-page/modules/ss-helpers/page.spec.js
@@ -43,7 +43,7 @@ describe('super/i-static-page/modules/ss-helpers/page', () => {
 			});
 
 			expect(fs.readFileSync(file).toString()).toBe(
-`
+				`
 Object.defineProperty(window, 'GLOBAL_NONCE', {
 \tvalue: undefined
 });