+++ ts-definitions/tests.d.ts
@@ -76,6 +76,50 @@ interface IncludeReturns extends Record<string, any> {
 declare function include<PATH extends keyof IncludeReturns>(path: PATH, ...args: any[]): IncludeReturns[PATH];
 
 declare namespace BrowserTests {
+	interface TestParams {
+		/**
+		 * Browser instance
+		 */
+		browser: import('playwright').Browser;
+
+		/**
+		 * Browser version
+		 */
+		version: string;
+
+		/**
+		 * Base test URL
+		 */
+		testURL: string;
+
+		/**
+		 * Initial page
+		 * @deprecated [better to create a new page for every spec](https://github.com/V4Fire/Client/tree/master/docs/tests#specs-isolation)
+		 */
+		page: Playwright.Page;
+
+		/**
+		 * Initial context
+		 * @deprecated [better to create a new context for every spec](https://github.com/V4Fire/Client/tree/master/docs/tests#specs-isolation)
+		 */
+		context: Playwright.BrowserContext;
+
+		/**
+		 * Context options
+		 */
+		contextOpts: Dictionary;
+
+		/**
+		 * Directory with component files
+		 */
+		componentDir: string;
+
+		/**
+		 * Directory to store some temporary data
+		 */
+		tmpDir: string;
+	}
+
 	/**
 	 * Helpers to write browser tests
 	 */
@@ -219,14 +263,32 @@ declare namespace BrowserTests {
 		 * @param page
 		 * @param context
 		 * @param [options]
+		 *
+		 * @deprecated
 		 */
 		setup(page: Playwright.Page, context: Playwright.BrowserContext, options?: SetupOptions): Promise<void>;
+
+		/**
+		 * Intercepts and collects all invoking of `console` methods on the specified page.
+		 * Mind, the intercepted callings aren't be shown a console till you invoke the `printPageLogs` method.
+		 *
+		 * @param page
+		 */
+		collectPageLogs(page: Playwright.Page);
+
+		/**
+		 * Prints all of the intercepted page console invokes to a console
+		 * @param page
+		 */
+		printPageLogs(page: Playwright.Page);
 
 		/**
 		 * Reloads the page and waits until `requestIdleCallback`
 		 *
 		 * @param page
 		 * @param [idleOptions]
+		 *
+		 * @deprecated
 		 */
 		reloadAndWaitForIdle(page: Playwright.Page, idleOptions?: WaitForIdleOptions): Promise<void>;
 
@@ -275,6 +337,8 @@ declare namespace BrowserTests {
 		 * @param componentSelector
 		 * @param props
 		 * @param [idleOptions]
+		 *
+		 * @deprecated
 		 */
 		setPropsToComponent(
 			page: Playwright.Page,
@@ -289,6 +353,8 @@ declare namespace BrowserTests {
 		 * @param page
 		 * @param componentSelector
 		 * @param props
+		 *
+		 * @deprecated
 		 */
 		reloadAndSetProps(
 			page: Playwright.Page,
@@ -459,6 +525,9 @@ declare namespace BrowserTests {
 		 * @param ctx
 		 * @param selector
 		 * @param [options]
+		 *
+		 * @deprecated
+		 * @see https://playwright.dev/docs/api/class-elementhandle#element-handle-wait-for-selector
 		 */
 		waitForEl(
 			ctx: PlaywrightElContext,
@@ -484,6 +553,11 @@ declare namespace BrowserTests {
 		 *
 		 * @param selectorOrElement
 		 * @param ctx
+		 *
+		 * @deprecated
+		 *
+		 * @see https://playwright.dev/docs/api/class-elementhandle#element-handle-is-visible
+		 * @see https://playwright.dev/docs/api/class-elementhandle#element-handle-wait-for-selector
 		 */
 		isVisible(selectorOrElement: PlaywrightElContext | string, ctx?: PlaywrightElContext): Promise<boolean>;
 
@@ -685,6 +759,8 @@ declare namespace BrowserTests {
 		 *
 		 * @param page
 		 * @param urls
+		 *
+		 * @deprecated
 		 */
 		waitForRequests(page: Playwright.Page, urls: string[]): Promise<void>;
 
@@ -694,6 +770,8 @@ declare namespace BrowserTests {
 		 *
 		 * @param page
 		 * @param urls
+		 *
+		 * @deprecated
 		 */
 		waitForRequestsFail(page: Playwright.Page, urls: string[]): Promise<void>;
 
@@ -704,6 +782,8 @@ declare namespace BrowserTests {
 		 * @param urls
 		 * @param response
 		 * @param timeout
+		 *
+		 * @deprecated
 		 */
 		interceptRequest(
 			page: Playwright.Page,
@@ -719,6 +799,8 @@ declare namespace BrowserTests {
 		 * @param urls
 		 * @param response
 		 * @param timeout
+		 *
+		 * @deprecated
 		 */
 		interceptRequests(
 			page: Playwright.Page,
@@ -733,6 +815,8 @@ declare namespace BrowserTests {
 		 * @param page
 		 * @param urls
 		 * @param event
+		 *
+		 * @deprecated
 		 */
 		waitForRequestsEvents(page: Playwright.Page, urls: string[], event: string): Promise<void>;
 
@@ -772,7 +856,6 @@ declare namespace Tests {
 	interface TestUtils {
 		/**
 		 * Returns a test that matches the `runner` parameter and the `name` passed to the `CLI`
-		 *
 		 * @param [options]
 		 */
 		getCurrentTest(options?: GetCurrentTestOptions): Function;